requirements.txt
anyio~=4.4.0
certifi~=2020.6.20
python-dateutil
exceptiongroup~=1.2.1
greenlet~=2.0.2
h11==0.14.0
httpcore~=1.0.5
httpx~=0.27.0
idna~=3.6
python-telegram-bot==21.6
sniffio==1.3.1
SQLAlchemy~=2.0.19
typing_extensions~=4.10.0
python-dotenv~=1.0.1
psycopg2-binary~=2.9.9
pytest~=8.2.0
pip~=24.0
attrs~=23.2.0
build~=1.0.3
wheel~=0.37.1
protobuf~=3.12.4
pyOpenSSL~=24.1.0
cryptography~=42.0.5
Pillow~=9.0.1
Jinja2~=3.1.4
keyring~=24.2.0
click~=8.1.6
psutil~=5.9.0
pytz~=2024.1
setuptools~=59.6.0
future~=0.18.2
importlib-metadata~=6.8.0
python-dateutil~=2.9.0.post0